Module modularise db
